Rename file to group_permission.dart and move file
