+++ app/lib/groups/src/pages/course/course_details/course_details_bloc.dart
@@ -12,7 +12,7 @@ import 'package:bloc_base/bloc_base.dart';
 import 'package:common_domain_models/common_domain_models.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:rxdart/subjects.dart';
-import 'package:sharezone/additional/course_permission.dart';
+import 'package:sharezone/groups/group_permission.dart';
 import 'package:sharezone/groups/src/models/splitted_member_list.dart';
 import 'package:sharezone/util/api/courseGateway.dart';
