+++ app/lib/groups/src/pages/school_class/card/school_class_card.dart
@@ -12,8 +12,8 @@ import 'package:design/design.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/additional/course_permission.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
+import 'package:sharezone/groups/group_permission.dart';
 import 'package:sharezone/groups/src/pages/school_class/edit/school_class_edit_page.dart';
 import 'package:sharezone/groups/src/widgets/group_share.dart';
 import 'package:sharezone_widgets/adaptive_dialog.dart';
