+++ app/lib/groups/src/pages/school_class/school_class_details.dart
@@ -10,7 +10,7 @@ import 'package:app_functions/app_functions.dart';
 import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/additional/course_permission.dart';
+import 'package:sharezone/groups/group_permission.dart';
 import 'package:sharezone/groups/src/pages/course/course_details/course_details_bloc.dart';
 import 'package:sharezone/groups/src/pages/course/course_details/course_settings.dart';
 import 'package:sharezone/groups/src/pages/course/course_details/write_permission_options.dart';
