+++ app/lib/timetable/timetable_permissions.dart
@@ -7,7 +7,7 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 import 'package:group_domain_models/group_domain_models.dart';
-import 'package:sharezone/additional/course_permission.dart';
+import 'package:sharezone/groups/group_permission.dart';
 
 /// User has to be a creator and a teacher or student
 bool hasPermissionToManageLessons(MemberRole myRole) {
